+++ client/my-sites/plans/plans.jsx
@@ -10,17 +10,16 @@ import { connect } from 'react-redux';
  */
 import config from 'config';
 import { getSelectedSiteId } from 'state/ui/selectors';
-import { getIntervalTypeFromCurrentPlan } from 'state/plans/selectors';
+import getIntervalTypeForTerm from 'state/plans/get-interval-type-for-term';
+import getCurrentPlanTerm from 'state/selectors/get-current-plan-term';
 import Plans from 'my-sites/plans/main';
 
-export default connect( ( state, ownProps ) => {
-	let intervalType = ownProps.intervalType;
-
+export default connect( ( state, { intervalType } ) => {
 	// For WP.com plans page, if intervalType is not explicitly specified in the URL,
 	// we want to show plans of the same term as plan that is currently active
 	if ( ! intervalType && config.isEnabled( 'upgrades/2-year-plans' ) ) {
 		const selectedSiteId = getSelectedSiteId( state );
-		intervalType = getIntervalTypeFromCurrentPlan( state, selectedSiteId );
+		intervalType = getIntervalTypeForTerm( getCurrentPlanTerm( state, selectedSiteId ) );
 	}
 
 	return {
