+++ client/state/plans/selectors.js
@@ -10,9 +10,7 @@ import { get, find } from 'lodash';
  * Internal Dependencies
  */
 import createSelector from 'lib/create-selector';
-import { calculateMonthlyPriceForPlan, getPlan as libGetPlan } from 'lib/plans';
-import { getSitePlan } from 'state/sites/selectors';
-import { TERM_TO_INTERVAL_TYPE } from './interval-types';
+import { calculateMonthlyPriceForPlan } from 'lib/plans';
 
 /**
  * Return WordPress plans getting from state object
@@ -85,17 +83,4 @@ export function getPlanSlug( state, productId ) {
 	const plan = getPlan( state, productId );
 
 	return get( plan, 'product_slug', null );
-}
-
-export function getIntervalTypeFromCurrentPlan( state, siteId ) {
-	const currentPlanProduct = getSitePlan( state, siteId );
-	if ( ! currentPlanProduct ) {
-		return null;
-	}
-
-	const currentPlanObject = libGetPlan( currentPlanProduct.product_slug );
-	if ( ! currentPlanObject ) {
-		return null;
-	}
-	return TERM_TO_INTERVAL_TYPE[ currentPlanObject.term ];
 }
